feat: Add Markdown renderers and common utilities for improved content rendering; implement specific renderers for document types
